Add normalizeEntityName to blueprints/ember-cli-sticky/index.js
I /think/ this fixes an error message when first running `ember install ember-cli-stick` that gives a message saying generate requires an entityname to be specified. 

```
The `ember generate` command requires an entity name to be specified. For more details, use `ember help'.
```

Also part of the ember-cli docs for addons that only add 3rd party libraries to the ember app.